Changed so that own and other profiles are accessed through profile/userId
