+++ src/app/[lang]/profile/[id]/page.tsx
@@ -1,41 +1,110 @@
+import ProfileView from '@/components/ProfileView';
 import Container from '@mui/material/Container/Container';
-import { getUserData } from '@/lib/prisma/users';
-import { Locale } from '@/lib/i18n/i18n-config';
-import { notFound } from 'next/navigation';
-import { isAdmin } from '@/lib/auth-utils';
-import UnauthorizedError from '@/components/UnauthorizedError';
+import CourseModal from '@/components/CourseModal';
 import { getServerAuthSession } from '@/lib/auth';
-
-import ProfileUserDetails from '@/components/ProfileView/ProfileUserDetails';
+import { notFound } from 'next/navigation';
+import { Locale } from '@i18n/i18n-config';
+import { translator } from '@/lib/i18n';
+import {
+  UserNamesAndIds,
+  getAllUsers,
+  getStudentNamesByCourseId,
+  getUserData,
+} from '@/lib/prisma/users';
+import {
+  RequestsAndUserNames,
+  getRequestsByCourseId,
+} from '@/lib/prisma/requests';
+import {
+  getTemplatesWithCreator,
+  getTemplatesByUserIdWithCreator,
+} from '@/lib/prisma/templates';
+import CreateTag from '@/app/[lang]/admin/dashboard/CreateTag';
+import CreateCountry from '@/app/[lang]/admin/dashboard/CreateCountry';
+import { getTags } from '@/lib/prisma/tags';
+import { getCountries } from '@/lib/prisma/country';
+import { isAdmin, isTrainerOrAdmin } from '@/lib/auth-utils';
+import UnauthorizedError from '@/components/UnauthorizedError';
 
 type Props = {
+  searchParams: { courseId?: string };
   params: {
     id: string;
     lang: Locale;
   };
 };
 
-export default async function ProfilePageById({ params }: Props) {
+export default async function ProfilePageById({ searchParams, params }: Props) {
   const session = await getServerAuthSession();
-  const userId = params.id;
-  const userData = await getUserData(userId);
+  const { t } = await translator(['components', 'admin']);
+  const allUsers = await getAllUsers();
+  const tags = await getTags();
+  const countries = await getCountries();
+
+  if (!isAdmin(session.user) && params.id !== session.user.id) {
+    return <UnauthorizedError lang={params.lang} />;
+  }
+
+  const userData =
+    isAdmin(session.user) && params.id !== session.user.id
+      ? await getUserData(params.id)
+      : await getUserData(session.user.id);
 
   if (!userData) {
     notFound();
   }
-  if (!isAdmin(session.user)) {
-    return <UnauthorizedError lang={params.lang} />;
+
+  const allCourses = userData?.createdCourses.concat(userData?.courses) ?? [];
+  const enrolledCourseIds = userData.courses.map((course) => course.id) ?? [];
+  const openedCourse = allCourses.find(
+    (course) => course.id === searchParams.courseId
+  );
+
+  const templates = isAdmin(session.user)
+    ? await getTemplatesWithCreator()
+    : await getTemplatesByUserIdWithCreator(session.user.id);
+
+  let enrolledStudents: UserNamesAndIds | null = [];
+  let requests: RequestsAndUserNames | null = [];
+  if (isTrainerOrAdmin(session.user) && openedCourse) {
+    enrolledStudents = await getStudentNamesByCourseId(openedCourse.id);
+    requests = await getRequestsByCourseId(openedCourse.id);
   }
 
   return (
     <Container maxWidth="md">
-      <div style={{ paddingBottom: '20px', paddingTop: '32px' }}>
-        <ProfileUserDetails
-          name={userData?.name ?? ''}
-          email={userData?.email ?? ''}
-          image={userData?.image ?? ''}
+      <CourseModal
+        lang={params.lang}
+        course={openedCourse}
+        usersEnrolledCourseIds={enrolledCourseIds}
+        enrolledStudents={enrolledStudents}
+        requests={requests}
+      />
+      <ProfileView
+        lang={params.lang}
+        userDetails={{
+          name: userData.name ?? '',
+          email: userData.email ?? '',
+          image: userData.image ?? '',
+        }}
+        courses={userData?.courses ?? []}
+        createdCourses={userData?.createdCourses ?? []}
+        users={allUsers}
+        templates={templates}
+        countries={countries}
+        tags={tags}
+      >
+        <CreateCountry
+          countryHeader={t('admin:CountriesSection.header')}
+          countries={countries}
+          lang={params.lang}
         />
-      </div>
+        <CreateTag
+          tagsHeader={t('admin:TagsSection.header')}
+          tags={tags}
+          lang={params.lang}
+        />
+      </ProfileView>
     </Container>
   );
 }
